move implementation into commonpaths.cpp
